AssayDesignerPanels jest to RTL
